+++ packages/client/src/utils/PushPipeline.ts
@@ -89,10 +89,6 @@ export class PushPipeline<InType, OutType = InType> extends Pipeline<InType, Out
         return this.source.length || 0
     }
 
-    isFull(): boolean {
-        return this.source.isFull()
-    }
-
     clear(): void {
         return this.source.clear()
     }