+++ packages/network/src/helpers/transformIterable.ts
@@ -1,5 +1,0 @@
-export const transformIterable = function *<F, T>(from: Iterable<F>, tranform: (f: F) => T): Iterable<T> {
-    for (const f of from) {
-        yield tranform(f)
-    }
-}